refactor VerifierAlgorithm test
